Bash script: Use specific Python interpreter
This fixes an issue when the `python` interpreter that's first in `$PATH` is
*not* the one that is using the `site-packages` where `mssql-cli` is installed.

This is especially an issue in an environment where there are lots of different
virtualenvs in `$PATH` (i.e. when using a separate virtualenv per installed
application) and the one where `mssql-cli` is installed is not the first one, so
the wrong interpreter would be used.

Actually, since the script should always be installed next to an interpreter
executable, the `else` part should not be neccessary, but I've kept it for
backwards compatibility with strange other setups.